chore: rename client to llm
