+++ tests/test_llm.py
@@ -1,4 +1,4 @@
-"""LightningLLM main tests."""
+"""LitAI main tests."""
 
 import os
 import re
@@ -10,15 +10,15 @@ from litai import LLM
 
 
 def test_initialization_with_config_file(monkeypatch):
-    """Test LigtningLLM config."""
+    """Test LitAI config."""
     mock_llm_instance = MagicMock()
-    monkeypatch.setattr("litai.client.SDKLLM", mock_llm_instance)
+    monkeypatch.setattr("litai.llm.SDKLLM", mock_llm_instance)
     LLM(model="openai/gpt-4", lightning_api_key="my-key", lightning_user_id="my-user-id")
     assert os.getenv("LIGHTNING_API_KEY") == "my-key"
     assert os.getenv("LIGHTNING_USER_ID") == "my-user-id"
 
 
-@patch("litai.client.SDKLLM")
+@patch("litai.llm.SDKLLM")
 def test_invalid_model(mock_llm_class):
     """Test invalid model name."""
     dummy_model_name = "dummy-model"
@@ -33,7 +33,7 @@ def test_invalid_model(mock_llm_class):
 def test_default_model(monkeypatch):
     """Test default model name."""
     mock_llm_instance = MagicMock()
-    monkeypatch.setattr("litai.client.SDKLLM", mock_llm_instance)
+    monkeypatch.setattr("litai.llm.SDKLLM", mock_llm_instance)
     warning_message = "No model was provided, defaulting to openai/gpt-4o"
     with pytest.warns(UserWarning, match=re.escape(warning_message)):
         llm = LLM()
@@ -41,7 +41,7 @@ def test_default_model(monkeypatch):
         assert llm.model == "openai/gpt-4o"
 
 
-@patch("litai.client.SDKLLM")
+@patch("litai.llm.SDKLLM")
 def test_cloudy_models_preload(mock_sdkllm):
     """Test that CLOUDY_MODELS are preloaded during LLM initialization."""
     cloudy_models = {
@@ -52,7 +52,7 @@ def test_cloudy_models_preload(mock_sdkllm):
         "google/gemini-2.5-pro",
         "google/gemini-2.5-flash",
     }
-    from litai.client import LLM as LLMCLIENT
+    from litai.llm import LLM as LLMCLIENT
 
     LLMCLIENT._sdkllm_cache.clear()
     llm = LLM()
@@ -67,10 +67,10 @@ def test_cloudy_models_preload(mock_sdkllm):
     assert set(enable_async_param) == {True, False}
 
 
-@patch("litai.client.SDKLLM")
+@patch("litai.llm.SDKLLM")
 def test_llm_context_length(mock_llm_class):
     """Test LigtningLLM context length."""
-    from litai.client import LLM as LLMCLIENT
+    from litai.llm import LLM as LLMCLIENT
 
     LLMCLIENT._sdkllm_cache.clear()
     mock_llm_instance = MagicMock()
@@ -83,10 +83,10 @@ def test_llm_context_length(mock_llm_class):
     assert llm.context_length() == 8000
 
 
-@patch("litai.client.SDKLLM")
+@patch("litai.llm.SDKLLM")
 def test_llm_chat(mock_llm_class):
     """Test LigtningLLM chat."""
-    from litai.client import LLM as LLMCLIENT
+    from litai.llm import LLM as LLMCLIENT
 
     LLMCLIENT._sdkllm_cache.clear()
     mock_llm_instance = MagicMock()
@@ -146,7 +146,7 @@ def test_model_override(monkeypatch):
             return mock_override
         raise ValueError(f"Unknown model: {name}")
 
-    monkeypatch.setattr("litai.client.SDKLLM", mock_llm_constructor)
+    monkeypatch.setattr("litai.llm.SDKLLM", mock_llm_constructor)
 
     llm = LLM(
         model="default-model",
@@ -175,7 +175,7 @@ def test_model_override(monkeypatch):
 
 def test_fallback_models(monkeypatch):
     """Test fallback model logic when main model fails."""
-    from litai.client import LLM as LLMCLIENT
+    from litai.llm import LLM as LLMCLIENT
 
     LLMCLIENT._sdkllm_cache.clear()
     mock_main_model = MagicMock()
@@ -197,7 +197,7 @@ def test_fallback_models(monkeypatch):
             return mock_fallback_model
         raise ValueError(f"Unknown model: {name}")
 
-    monkeypatch.setattr("litai.client.SDKLLM", mock_llm_constructor)
+    monkeypatch.setattr("litai.llm.SDKLLM", mock_llm_constructor)
 
     llm = LLM(
         model="main-model",
@@ -229,7 +229,7 @@ async def test_llm_async_chat(monkeypatch):
     mock_sdkllm.name = "mock-model"
     mock_sdkllm.chat = AsyncMock(return_value="Hello, async world!")
 
-    monkeypatch.setattr("litai.client.SDKLLM", lambda *args, **kwargs: mock_sdkllm)
+    monkeypatch.setattr("litai.llm.SDKLLM", lambda *args, **kwargs: mock_sdkllm)
 
     llm = LLM(model="mock-model", enable_async=True)
     result = await llm.chat("Hi there", conversation="async-test")
@@ -248,7 +248,7 @@ def test_get_history(monkeypatch, capsys):
         ]
     )
 
-    monkeypatch.setattr("litai.client.SDKLLM", lambda *args, **kwargs: mock_sdkllm)
+    monkeypatch.setattr("litai.llm.SDKLLM", lambda *args, **kwargs: mock_sdkllm)
 
     llm = LLM(model="mock-model")
 
@@ -282,7 +282,7 @@ def test_authenticate_method(monkeypatch):
     def mock_auth_constructor():
         return mock_auth
 
-    monkeypatch.setattr("litai.client.login.Auth", mock_auth_constructor)
+    monkeypatch.setattr("litai.llm.login.Auth", mock_auth_constructor)
 
     # Test case 1: Both api_key and user_id provided
     LLM(model="openai/gpt-4", lightning_api_key="my-key", lightning_user_id="my-user-id")
@@ -305,10 +305,10 @@ def test_authenticate_method(monkeypatch):
     mock_auth.authenticate.assert_called_once()
 
 
-@patch("litai.client.SDKLLM")
+@patch("litai.llm.SDKLLM")
 def test_llm_if_method(mock_sdkllm_class):
     """Test the LLM if_ method."""
-    from litai.client import LLM as LLMCLIENT
+    from litai.llm import LLM as LLMCLIENT
 
     LLMCLIENT._sdkllm_cache.clear()
 
@@ -329,10 +329,10 @@ def test_llm_if_method(mock_sdkllm_class):
     assert llm.if_("the product is amazing", "is it a positive response?") is True
 
 
-@patch("litai.client.SDKLLM")
+@patch("litai.llm.SDKLLM")
 def test_llm_classify_method(mock_sdkllm_class):
     """Test the LLM classify method."""
-    from litai.client import LLM as LLMCLIENT
+    from litai.llm import LLM as LLMCLIENT
 
     LLMCLIENT._sdkllm_cache.clear()
